Quantity changes

+++ app/controllers/orders_controller.rb
@@ -9,15 +9,36 @@ class OrdersController < ApplicationController
   def checkout
     userid = current_user.id
     @orders =  CartItem.where("user_id = #{userid}")
+    flag = false
     @orders.each do |order|
-      name = order.name
-      description = order.description
-      price = order.price
-      quantity = order.quantity
-      itemid = order.item_id
-      status = "Placed"
-      Order.create(:name=>name, :description=>description, :price=>price, :quantity=>quantity, :status=>status, :user_id=>userid, :item_id=>itemid)
-      order.destroy
+      if order.quantity > order.available
+        flag = true
+        break
+      end
+    end
+    if flag
+      flash[:notice] = 'Sorry, some products are not available please edit your cart.'
+      redirect_to cart_items_path
+    else
+      @orders.each do |order|
+        name = order.name
+        description = order.description
+        price = order.price
+        quantity = order.quantity
+        itemid = order.item_id
+        status = "Placed"
+        Order.create(:name=>name, :description=>description, :price=>price, :quantity=>quantity, :status=>status, :user_id=>userid, :item_id=>itemid)
+        order.destroy
+        @item = Item.find(itemid)
+        quantity = @item.quantity - quantity
+        @item.quantity = quantity
+        @item.save
+        @cartitems =  CartItem.where("item_id = #{itemid}")
+        @cartitems.each do |cartitem|
+          cartitem.available = quantity
+          cartitem.save
+        end
+      end
     end
   end
 
@@ -30,8 +51,18 @@ class OrdersController < ApplicationController
 
   def destroy
     @order = Order.find(params[:id])
+    itemid = @order.item_id
+    quantity = @order.quantity
     @order.destroy
-
+    @item = Item.find(itemid)
+    quantity = @item.quantity + quantity
+    @item.quantity = quantity
+    @item.save
+    @cartitems =  CartItem.where("item_id = #{itemid}")
+    @cartitems.each do |cartitem|
+      cartitem.available = quantity
+      cartitem.save
+    end
     redirect_to orders_path
   end
